updated logic and tests
